FAIR data pipeline manual

This is the updated manual for the upcoming FAIR interfaced version of the Data Registry package.

FAIR data pipeline manual

Managing code runs

Reading data

Writing to the Data Registry

Registering data

Raising issues

What's my file?

Index

Note that data products are processed and downloaded at the point of initialisation, provided that a data_config file is specified, and the offline_mode option is not used.

Managing code runs

DataPipeline.initialise — Function

initialise(config_file, submission_script)

Read [working] config.yaml file. Returns a DataRegistryHandle containing:

the working config.yaml file contents

the object id for this file

the object id for the submission script file

DataPipeline.finalise — Function

finalise(handle)

Complete (i.e. finish) code run.

Reading data

DataPipeline.read_array — Function

read_array(db::SQLite.DB, data_product::String[, component::String]; ... )

Read HDF5 array data resource(s) – optionally specify the individual component.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

data_product – data product search string, e.g. "human/infection/SARS-CoV-2/".

component – [optionally] specify the component name.

Options

version – optionally specify the version (else the latest version is read.)

fuzzy_match – set false for exact matches only.

flatten – set true for a flattened 2d table.

use_axis_arrays – set true to return the matching data as AxisArray types.

log_access – set false to supress data access logging.

data_log_id – (optionally) specify the data log id.

read_array(handle, data_product, [component])

Read [array] data product.

note that it must already have been downloaded from the remote data store using fdp pull.

the latest version of the data is read unless otherwise specified.

DataPipeline.read_table — Function

read_table(db::SQLite.DB, data_product::String, component::String)

SQLite Data Registry helper function. Search and return [HDF5] table data as a DataFrame.

Parameters

db – SQLite.DB object.

data_product – data product search string, e.g. "human/infection/SARS-CoV-2/".

component – as above, [required] search string for components names.

Options

version – optionally specify the version (else the latest version is read.)

fuzzy_match – set false for exact matches only.

log_access – set false to supress data access logging.

data_log_id – (optionally) specify the data log id.

read_table(handle, data_product; component, version)

Read [table] data product.

note that it must already have been downloaded from the remote data store using fdp pull.

the latest version of the data is read unless otherwise specified.

DataPipeline.read_estimate — Function

read_estimate(cn::SQLite.DB, data_product::String, [component::String]; ... )

Read key-value pair, e.g. a point-estimate.

SQLite Data Registry helper function. Search TOML-based data resources stored in cn, a SQLite database created previously by a call to fetch_data_per_yaml.

Parameters

cn – SQLite.DB object.

data_product – data product search string, e.g. "human/infection/SARS-CoV-2/".

component – as above, [optional] search string for components names.

Options

version – optionally specify the version (else the latest version is read.)

fuzzy_match – set false for exact matches only.

key – specify to return .toml keys of a particular type, e.g. "type" or "value".

data_type – specify to return an array of this type instead of a DataFrame.

log_access – set false to supress data access logging.

data_log_id – (optionally) specify the data log id.

read_estimate(handle, data_product, [component])

Read TOML-based data product.

note that it must already have been downloaded from the remote data store using fdp pull.

the specific version can be specified in the config file (else the latest version is used.)

DataPipeline.read_distribution — Function

readdistribution(handle, dataproduct, [component])

Read TOML-based data product.

note that it must already have been downloaded from the remote data store using fdp pull.

the specific version can be specified in the config file (else the latest version is used.)

DataPipeline.link_read — Function

link_read(handle, data_product)

Returns the file path of a data product that has been registered in the local data registry, either directly or via the CLI.

Writing to the Data Registry

The process of registering objects such as data, code, and model runs, in the main Data Registry involves two steps; [local] registration, and then committing registered objects to the main online Registry.

Registering data

DataPipeline.write_array — Function

write_array(handle, data, data_product, component)

Write an array as a component to an hdf5 file.

WIP.

DataPipeline.write_table — Function

write_table(handle, table, data_product)

Write a Tables.jl interface input (https://github.com/JuliaData/Tables.jl) to file (and register a corresonding data product in the local Data Registry.)

DataPipeline.write_estimate — Function

write_estimate(handle, value, data_product, component)

Write point estimate to file (and register a corresonding data product in the local Data Registry.)

DataPipeline.write_distribution — Function

write_distribution(handle, distribution, parameters, data_product, component)

Write specification of a statistical distribution to file (and register a corresonding data product in the local Data Registry.)

DataPipeline.link_write — Function

link_write(handle, filepath, data_product)

Registers a file-based data product based on information provided in the working config file, e.g. for writing external objects.

Raising issues

DataPipeline.raise_issue — Function

raise_issue(handle; ... )

Register issue with data product; component; external object; or script.

Pass the object URI as a named parameter[s], e.g. raise_issue(handle; data_product=dp, component=comp).

Optional parameters

data_product

component

external_object

script

What's my file?

DataPipeline.whats_my_file — Function

whats_my_file(path::String; show_path=false)

Search the Data Registry for matches with a given local file (or directory of files.)

Parameters

path – local file path, or directory.

show_path – (optional) display the [remote] path of any matching files.

DataPipeline.registry_audit — Function

registry_audit(url; trace="both")

Search the Data Registry for known issues with, e.g. data products, code repo releases or code runs.

Any issues that impact upon provenance (i.e. [a subset of] the graph of Registry objects associated with this one) can also be displayed w.r.t. "inputs", "outputs" or "both".

Parameters

url – the URL of e.g. a data product or code repo release in the Data Registry.

trace – "inputs", "outputs" or "both" – also the default.

Index

DataPipeline.finalise

DataPipeline.initialise

DataPipeline.link_read

DataPipeline.link_write

DataPipeline.raise_issue

DataPipeline.read_array

DataPipeline.read_distribution

DataPipeline.read_estimate

DataPipeline.read_table

DataPipeline.registry_audit

DataPipeline.whats_my_file

DataPipeline.write_array

DataPipeline.write_distribution

DataPipeline.write_estimate

DataPipeline.write_table