FAIR data pipeline manual

This is the updated manual for the upcoming FAIR interfaced version of the Data Registry package.

FAIR data pipeline manual

Managing code runs

Reading data

Writing to the Data Registry

Registering data

Raising issues

What's my file?

Index

Note that data products are processed and downloaded at the point of initialisation, provided that a data_config file is specified, and the offline_mode option is not used.

Managing code runs

DataPipeline.initialise — Function

initialise(config_file, submission_script)

Reads in working config.yaml file, generates a new Code Run entry, and returns a DataRegistryHandle containing various metadata.

DataPipeline.finalise — Function

finalise(handle)

Complete (i.e. finish) code run.

Reading data

DataPipeline.read_array — Function

read_array(handle, data_product[, component])

Read [array] data product.

note that it must already have been downloaded from the remote data store using fdp pull.

the latest version of the data is read unless otherwise specified.

DataPipeline.read_table — Function

read_table(handle, data_product[, component])

Read [table] data product.

note that it must already have been downloaded from the remote data store using fdp pull.

the latest version of the data is read unless otherwise specified.

DataPipeline.read_estimate — Function

read_estimate(handle, data_product, [component])

Read TOML-based data product.

note that it must already have been downloaded from the remote data store using fdp pull.

the specific version can be specified in the config file (else the latest version is used.)

DataPipeline.read_distribution — Function

readdistribution(handle, dataproduct, [component])

Read TOML-based data product.

note that it must already have been downloaded from the remote data store using fdp pull.

the specific version can be specified in the config file (else the latest version is used.)

Missing docstring.

Missing docstring for link_read. Check Documenter's build log for details.

Writing to the Data Registry

The process of registering objects such as data, code, and model runs, in the main Data Registry involves two steps; [local] registration, and then committing registered objects to the main online Registry.

Registering data

DataPipeline.write_array — Function

write_array(handle, data, data_product, component)

Write an array as a component to an hdf5 file.

See also: write_table, read_array, read_table

DataPipeline.write_table — Function

write_table(handle, data, data_product, component)

Write a table as a component to an hdf5 file.

See also: write_array, read_array, read_table

DataPipeline.write_estimate — Function

write_estimate(handle, value, data_product, component)

Write a point estimate as a component to a toml file.

DataPipeline.write_distribution — Function

write_distribution(handle, distribution, parameters, data_product, component)

Write a distribution as a component to a toml file.

Missing docstring.

Missing docstring for link_write. Check Documenter's build log for details.

Raising issues

DataPipeline.raise_issue — Function

raise_issue(handle; ... )

Register issue with data product; component; external object; or script.

Pass the object URI as a named parameter[s], e.g. raise_issue(handle; data_product=dp, component=comp).

Optional parameters

data_product

component

external_object

script

What's my file?

Missing docstring.

Missing docstring for whats_my_file. Check Documenter's build log for details.

Missing docstring.

Missing docstring for registry_audit. Check Documenter's build log for details.

Index

DataPipeline.finalise

DataPipeline.initialise

DataPipeline.raise_issue

DataPipeline.read_array

DataPipeline.read_distribution

DataPipeline.read_estimate

DataPipeline.read_table

DataPipeline.write_array

DataPipeline.write_distribution

DataPipeline.write_estimate

DataPipeline.write_table