FDP interface manual

FDP interface manual

Reading / downloading data

Downloading data products

Reading data

Extenal objects

Writing to the Data Registry

Registering objects

Raising issues

Other

Index

Note that data products are processed and downloaded at the point of initialisation, provided that a data_config file is specified, and the offline_mode option is not used.

Reading / downloading data

Downloading data products

DataRegistryUtils.initialise — Function

initialise(config_file)

Read working config.yaml file. Returns a DataRegistryHandle containing:

the working config.yaml file contents

the object id for this file

the object id for the submission script file

Reading data

Note that the $fdp_$ prefix is temporary.

DataRegistryUtils.fdp_read_estimate — Function

fdp_read_estimate(handle, data_product; component, version)

Read TOML-based data product.

note that it must already have been downloaded from the remote data store using fdp pull.

the latest version of the data is read unless otherwise specified.

DataRegistryUtils.fdp_read_array — Function

fdp_read_array(handle, data_product; component, version)

Read [array] data product.

note that it must already have been downloaded from the remote data store using fdp pull.

the latest version of the data is read unless otherwise specified.

DataRegistryUtils.fdp_read_table — Function

fdp_read_table(handle, data_product; component, version)

Read [table] data product.

note that it must already have been downloaded from the remote data store using fdp pull.

the latest version of the data is read unless otherwise specified.

Extenal objects

NB. per the [Data pipeline docs][dpdocs], use linkread() and linkwrite() to read and write external objects (rather than the standard API ``readxxx$and$write_xxx`` functions.)

DataRegistryUtils.link_read — Function

link_read(handle, alias)

This function returns the path of an external object in the local data store:

If the alias is already recorded in the handle, returns the path. If not, find the location of the file referenced by its alias.

Also, stores metadata associated with the external object.

Note that the alias is not recorded in the data registry. Rather, it is a means to reference external objects in the config.yaml file.

DataRegistryUtils.link_write — Function

link_write(handle, alias)

For writing external objects.

Use linkread() and linkwrite() to read and write external objects, rather than the standard API readxxx() and writexxx() calls.

Writing to the Data Registry

The process of registering objects such as data, code, and model runs, in the main Data Registry involves two steps; [local] registration, and then committing registered objects to the main online Registry.

Registering objects

DataRegistryUtils.write_array — Function

write_array()

Write an array as a component to an hdf5 file.

WIP.

WIP. TBA:

estimate, table

model

model run? or is this done only via scripts?

Raising issues

DataRegistryUtils.raise_issue — Function

raise_issue(handle; ... )

Register issue with data product; component; external object; or script.

Pass the object URI as a named parameter[s], e.g. $raise_issue(handle; data_product=dp, component=comp)$.

Optional parameters

data_product

component

external_object

script

Other

DataRegistryUtils.whats_my_file — Function

whats_my_file(path::String; show_path=false)

Search the Data Registry for matches with a given local file (or directory of files.)

Parameters

path – local file path, or directory.

show_path – (optional) display the [remote] path of any matching files.

DataRegistryUtils.registry_audit — Function

registry_audit(url; trace="both")

Search the Data Registry for known issues with, e.g. data products, code repo releases or code runs.

Any issues that impact upon provenance (i.e. [a subset of] the graph of Registry objects associated with this one) can also be displayed w.r.t. "inputs", "outputs" or "both".

Parameters

url – the URL of e.g. a data product or code repo release in the Data Registry.

trace – "inputs", "outputs" or "both" – also the default.

Index

DataRegistryUtils.commit_all

DataRegistryUtils.commit_staged_data_product

DataRegistryUtils.commit_staged_model

DataRegistryUtils.commit_staged_run

DataRegistryUtils.fdp_read_array

DataRegistryUtils.fdp_read_estimate

DataRegistryUtils.fdp_read_table

DataRegistryUtils.initialise

DataRegistryUtils.initialise_local_registry

DataRegistryUtils.link_read

DataRegistryUtils.link_write

DataRegistryUtils.load_array!

DataRegistryUtils.raise_issue

DataRegistryUtils.read_array

DataRegistryUtils.read_data_product_from_file

DataRegistryUtils.read_estimate

DataRegistryUtils.read_table

DataRegistryUtils.register_data_product

DataRegistryUtils.register_github_model

DataRegistryUtils.register_model_run

DataRegistryUtils.registry_audit

DataRegistryUtils.registry_commit_status

DataRegistryUtils.whats_my_file

DataRegistryUtils.write_array

[dpdocs]: https://scottishcovidresponse.github.io/docs/datapipeline/interface/