Package manual

Package manual

Reading / downloading data

Process data products

Reading data

Writing to the Data Registry

Register model

Register model run

Other

Index

Reading / downloading data

Process data products

DataRegistryUtils.fetch_data_per_yaml — Function

fetch_data_per_yaml(yaml_filepath, out_dir = "./out/"; ... )

Refresh and load data products from the SCRC data registry.

Checks the file hash for each data product and downloads anew any that are determined to be out-of-date. Allowing that the function has been called previously (i.e. so that the data has already been downloaded,)

Parameters

yaml_filepath – the location of a .yaml file.

out_dir – the local system directory where data will be stored.

Named parameters

offline_mode – set true when the Data Registry's RESTful API is inaccessible, e.g. no internet connection.

sql_file – (optional) SQL file for e.g. custom SQLite views, indexes, or whatever.

db_path – (optionally) specify the filepath of the database to use (or create.)

force_db_refresh – set true to overide file hash check on database inserts.

auto_logging – set true to enable automatic data access logging.

verbose – set to true to show extra output in the console.

Reading data

DataRegistryUtils.read_estimate — Function

read_estimate(cn::SQLite.DB, data_product::String, [component::String]; data_type=nothing)

Read key-value pair, e.g. a point-estimate.

SQLite Data Registry helper function. Search TOML-based data resources stored in cn, a SQLite database created previously by a call to fetch_data_per_yaml.

Parameters

cn – SQLite.DB object.

data_product – data product search string, e.g. "human/infection/SARS-CoV-2/%".

component – as above, [optional] search string for components names.

key – (optionally) specify to return .toml keys of a particular type, e.g. "type" or "value".

data_type – (optionally) specify to return an array of this type instead of a DataFrame.

log_access – set false to supress data access logging.

data_log_id – (optionally) specify the data log id.

Missing docstring.

Missing docstring for read_array. Check Documenter's build log for details.

DataRegistryUtils.read_table — Function

read_table(cn::SQLite.DB, data_product::String, component::String)

SQLite Data Registry helper function. Search and return [HDF5] table data as a DataFrame.

Parameters

cn – SQLite.DB object.

data_product – data product search string, e.g. "human/infection/SARS-CoV-2/%".

component – as above, [required] search string for components names.

DataRegistryUtils.read_data_product_from_file — Function

read_data_product_from_file(filepath; use_axis_arrays = false, verbose = false)

Read HDF5 or TOML file from local system.

Parameters

filepath – the location of an HDF5 or TOML file.

use_axis_arrays – convert the output to AxisArrays, where applicable.

verbose – set to true to show extra output in the console.

Writing to the Data Registry

Register model

DataRegistryUtils.register_github_model — Function

register_github_model(model_config, scrc_access_tkn; ... )
register_github_model(db, model_config, scrc_access_tkn; ... )
register_github_model(db, model_name, model_version, model_repo, scrc_access_tkn; ... )

Stage model code as a code_repo_release for upload the SCRC data registry.

The staged code_repo_release can then be pushed to the main Data Registry using e.g. the commit_staged_model function. If used, the model_config file should include (at a minimum) the $model_name$, $model_version$ and $model_repo$ fields. Else these can be passed directly to the function.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

model_config – path to the model config .yaml file.

model_name – label for the model release.

model_version – version number in the format 'n.n.n', e.g. $0.0.1$.

model_repo – url of the model [e.g. GitHub] repo.

model_hash – arbitrary identifying 'hash' string of the model.

Named parameters

model_description – (optional) description of the model.

model_website – (optional) website, e.g. for an accompanying paper, blog, or model documentation – model_repo by default.

storage_root_url – E.g. "https://github.com/" – also the default.

storage_root_id – Data Registry storage root identifier, "https://data.scrc.uk/api/storage_root/11/" by default.

Missing docstring.

Missing docstring for commit_staged_model. Check Documenter's build log for details.

Register model run

DataRegistryUtils.register_model_run — Function

register_model_run(db, model_staging_id, submission_script_text, model_run_description)

Stage a model run for subsequent upload to the $code_run$ endpoint of the SCRC Data Registry.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

model_id – [Internal] model id.

model_config – path to the model config .yaml file.

submission_script_text – e.g. 'julia my/julia/code.jl'.

model_run_description – description of the model run.

Missing docstring.

Missing docstring for commit_staged_run. Check Documenter's build log for details.

Other

Missing docstring.

Missing docstring for register_text_file. Check Documenter's build log for details.

Index

DataRegistryUtils.fetch_data_per_yaml

DataRegistryUtils.read_data_product_from_file

DataRegistryUtils.read_estimate

DataRegistryUtils.read_table

DataRegistryUtils.register_github_model

DataRegistryUtils.register_model_run