Package manual

Package manual

Reading / downloading data

Downloading data products

Reading data

Writing to the Data Registry

Registering objects locally

Committing objects to the online Registry

Other

Index

Reading / downloading data

Downloading data products

Note that data products are processed and downloaded at the point of initialisation, provided that a data_config file is specified, and the offline_mode option is not used.

DataRegistryUtils.initialise_local_registry — Function

initialise_local_registry(out_dir = "./out/"; ... )

Refresh and load data products from the SCRC data registry.

Checks the file hash for each data product and downloads anew any that are determined to be out-of-date. Allowing that the function has been called previously (i.e. so that the data has already been downloaded,)

Parameters

out_dir – the local system directory where data will be stored.

data_config – the location of the .yaml config file (see the docs.)

offline_mode – set true when the Data Registry's RESTful API is inaccessible, e.g. no internet connection.

sql_file – (optional) SQL file for e.g. custom SQLite views, indexes, or whatever.

db_path – (optionally) specify the filepath of the database to use (or create.)

auto_logging – set true to enable automatic data access logging.

verbose – set true to show extra output in the console.

force_db_refresh – set true to overide file hash check on local refresh, debugging only (hint: in the event of corrupted data, it would be easier to clear the designated directory of all data files and start anew.)

Reading data

Note that arrays can be processed into flat 2d tables, using either load_array! or read_array with the flatten=true.

Missing docstring.

Missing docstring for read_estimate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for read_array. Check Documenter's build log for details.

DataRegistryUtils.load_array! — Function

load_array!(db::SQLite.DB, data_product::String[, component::String]; ... )

Load HDF5 array(s) data resource.

SQLite Data Registry helper function. Optionally specify the individual component.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

data_product – data product search string, e.g. "human/infection/SARS-CoV-2/".

component – [optionally] specify the component name.

Options

version – optionally specify the version (else the latest version is read.)

fuzzy_match – set false for exact matches only.

sql_alias – optional alias for the database table.

log_access – set false to supress data access logging.

data_log_id – (optionally) specify the data log id.

Missing docstring.

Missing docstring for read_table. Check Documenter's build log for details.

DataRegistryUtils.read_data_product_from_file — Function

read_data_product_from_file(filepath; use_axis_arrays = false, verbose = false)

Read HDF5, CSV or TOML file from local system.

Parameters

filepath – the location of an e.g. HDF5 file.

use_axis_arrays – convert the output to AxisArrays, where applicable.

verbose – set to true to show extra output in the console.

Writing to the Data Registry

The process of registering objects such as data, code, and model runs, in the main Data Registry involves two steps; [local] registration, and then committing registered objects to the main online Registry.

Registering objects locally

DataRegistryUtils.register_data_product — Function

register_data_product(db, dp_name, version, filepath; namespace="SCRC")

Stage a data set (i.e. a local file) for upload to the $data_product$ endpoint of the SCRC Data Registry.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

dp_name – data product name.

version – e.g. "0.0.1"

filepath – path of the e.g. HDF5 file.

namespace – (optionally) specify the namespace, "SCRC" by default.

storage_root_id – Data Registry storage root identifier, "https://data.scrc.uk/api/storage_root/11/" (GitHub) by default.

DataRegistryUtils.register_github_model — Function

register_github_model(model_config, scrc_access_tkn; ... )
register_github_model(db, model_config, scrc_access_tkn; ... )
register_github_model(db, model_name, model_version, model_repo, scrc_access_tkn; ... )

Stage model code as a code_repo_release for upload the SCRC data registry.

The staged code_repo_release can then be pushed to the main Data Registry using e.g. the commit_staged_model function. If used, the model_config file should include (at a minimum) the $model_name$, $model_version$ and $model_repo$ fields. Else these can be passed directly to the function.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

model_config – path to the model config .yaml file.

model_name – label for the model release.

model_version – version number in the format 'n.n.n', e.g. $0.0.1$.

model_repo – url of the model [e.g. GitHub] repo.

model_hash – arbitrary identifying 'hash' string of the model.

Named parameters

model_description – (optional) description of the model.

model_website – (optional) website, e.g. for an accompanying paper, blog, or model documentation – model_repo by default.

storage_root_id – Data Registry storage root identifier, "https://data.scrc.uk/api/storage_root/11/" (GitHub) by default.

DataRegistryUtils.register_model_run — Function

register_model_run(db, model_id, model_config, submission_script_text, model_run_description)

Stage a model run for subsequent upload to the $code_run$ endpoint of the SCRC Data Registry.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

model_id – [Internal] model id.

model_config – path to the model config .yaml file.

submission_script_text – e.g. 'julia my/julia/code.jl'.

model_run_description – description of the model run.

Committing objects to the online Registry

Note that 'staged' objects (i.e. registered locally) can be committed all at once using commit_all, or one at a time using the identifiers yielded by the above function calls, e.g. register_data_product.

DataRegistryUtils.registry_commit_status — Function

registry_commit_status(db)

Display the current commit status of the local staging Registry.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

DataRegistryUtils.commit_all — Function

commit_all(db, scrc_access_tkn)

Commit all outstanding (i.e. 'staged') objects to the Data Registry.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

scrc_access_tkn – access token (see https://data.scrc.uk/docs/.)

ftp_username – optionally specify

ftp_password – optionally specify

DataRegistryUtils.commit_staged_data_product — Function

commit_staged_data_product(db, staging_id, scrc_access_tkn)

Commit staged [local] file to the $data_product$ endpoint of the SCRC Data Registry.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

staging_id – data product id, obtained when the model was [pre-]registered.

scrc_access_tkn – access token (see https://data.scrc.uk/docs/.)

DataRegistryUtils.commit_staged_model — Function

commit_staged_model(db, staging_id, scrc_access_tkn)

Upload model run to the $code_run$ endpoint of the SCRC Data Registry.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

staging_id – model id, obtained when the model was [pre-]registered.

scrc_access_tkn – access token (see https://data.scrc.uk/docs/.)

DataRegistryUtils.commit_staged_run — Function

commit_staged_run(db, staging_id, scrc_access_tkn)

Upload model run to the $code_run$ endpoint of the SCRC Data Registry.

Parameters

db – SQLite.DB object yielded by previous call to fetch_data_per_yaml.

staging_id – run id, obtained when the run was [pre-]registered.

scrc_access_tkn – access token (see https://data.scrc.uk/docs/.)

Other

Missing docstring.

Missing docstring for whats_my_file. Check Documenter's build log for details.

Missing docstring.

Missing docstring for registry_audit. Check Documenter's build log for details.

Index

DataRegistryUtils.commit_all

DataRegistryUtils.commit_staged_data_product

DataRegistryUtils.commit_staged_model

DataRegistryUtils.commit_staged_run

DataRegistryUtils.finalise

DataRegistryUtils.initialise

DataRegistryUtils.initialise_local_registry

DataRegistryUtils.link_read

DataRegistryUtils.link_write

DataRegistryUtils.load_array!

DataRegistryUtils.raise_issue

DataRegistryUtils.read_array

DataRegistryUtils.read_data_product_from_file

DataRegistryUtils.read_distribution

DataRegistryUtils.read_estimate

DataRegistryUtils.read_table

DataRegistryUtils.register_data_product

DataRegistryUtils.register_github_model

DataRegistryUtils.register_model_run

DataRegistryUtils.registry_audit

DataRegistryUtils.registry_commit_status

DataRegistryUtils.whats_my_file

DataRegistryUtils.write_array

DataRegistryUtils.write_distribution

DataRegistryUtils.write_estimate

DataRegistryUtils.write_table