Package manual

Package manual

Reading / downloading data

Downloading data products

Reading data

Writing to the Data Registry

Registering objects locally

Committing objects to the online Registry

Other

Index

Reading / downloading data

Downloading data products

Note that data products are processed and downloaded at the point of initialisation, provided that a data_config file is specified, and the offline_mode option is not used.

Missing docstring.

Missing docstring for initialise_local_registry. Check Documenter's build log for details.

Reading data

Note that arrays can be processed into flat 2d tables, using either load_array! or read_array with the flatten=true.

Missing docstring.

Missing docstring for read_estimate. Check Documenter's build log for details.

Missing docstring.

Missing docstring for read_array. Check Documenter's build log for details.

Missing docstring.

Missing docstring for load_array!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for read_table. Check Documenter's build log for details.

Missing docstring.

Missing docstring for read_data_product_from_file. Check Documenter's build log for details.

Writing to the Data Registry

The process of registering objects such as data, code, and model runs, in the main Data Registry involves two steps; [local] registration, and then committing registered objects to the main online Registry.

Registering objects locally

DataPipeline.register_data_product — Function

register_data_product(handle, data_product)

Register data product (from link_write())

Missing docstring.

Missing docstring for register_github_model. Check Documenter's build log for details.

Missing docstring.

Missing docstring for register_model_run. Check Documenter's build log for details.

Committing objects to the online Registry

Note that 'staged' objects (i.e. registered locally) can be committed all at once using commit_all, or one at a time using the identifiers yielded by the above function calls, e.g. register_data_product.

Missing docstring.

Missing docstring for registry_commit_status. Check Documenter's build log for details.

Missing docstring.

Missing docstring for commit_all. Check Documenter's build log for details.

Missing docstring.

Missing docstring for commit_staged_data_product. Check Documenter's build log for details.

Missing docstring.

Missing docstring for commit_staged_model. Check Documenter's build log for details.

Missing docstring.

Missing docstring for commit_staged_run. Check Documenter's build log for details.

Other

Missing docstring.

Missing docstring for whats_my_file. Check Documenter's build log for details.

Missing docstring.

Missing docstring for registry_audit. Check Documenter's build log for details.

Index

DataPipeline.finalise

DataPipeline.initialise

DataPipeline.link_read

DataPipeline.link_write

DataPipeline.raise_issue

DataPipeline.read_array

DataPipeline.read_distribution

DataPipeline.read_estimate

DataPipeline.read_table

DataPipeline.register_data_product

DataPipeline.write_array

DataPipeline.write_distribution

DataPipeline.write_estimate

DataPipeline.write_table